% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kfit.R
\name{kfit}
\alias{kfit}
\title{kfit}
\usage{
kfit(df, x = Index, y, b = 0.1, w = 1, k = "Nm")
}
\arguments{
\item{df}{A dataframe that will be processed.}

\item{x}{Numeric vector that is the same length as y (default is Index column generated by qclean function).}

\item{y}{Numeric vector that is the same length as x.}

\item{b}{The bandwidth of the kernel (set to 0.1 by default).}

\item{w}{Weight vector for kernel estimation.}

\item{k}{Character string to indicate kernel function (default is "Nm" for normal kernel)
"Ep" for Epanechnikov or "Bw" for Biweight are alternative options}
}
\value{
Returns a dataframe that contains a column with kernel estimates.
}
\description{
kfit creates a kernel estimate using the Nadaraya-Watson kernel estimator.
}
