% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ma.R
\name{ma}
\alias{ma}
\title{ma}
\usage{
ma(df, x, y, alpha = 1, n = 7, sma = TRUE, ema = TRUE, date = TRUE, u = "days")
}
\arguments{
\item{df}{A dataframe that will be processed.}

\item{x}{The variable to average y over.}

\item{y}{The variable to be averaged.}

\item{alpha}{Smoothing parameter for the exponential moving average. A higher alpha places more emphasis on the current observation while
a lower alpha places more emphasis on previous observations (set to 1 by defauilt).}

\item{n}{Number of observations taken into account by the moving average algorithms (set to 7 by default).}

\item{sma}{If sma is true, calculates the simple moving average and stores it in a variable named "sma" (set to TRUE by default).}

\item{ema}{If ema is true, calculates the exponential moving average and stores it in a variable named "ema" (set to TRUE by default).}

\item{date}{If date is TRUE, converts x variable to Date for calculating moving averages (set to TRUE by default).}

\item{u}{Units for the x variable to calculate time interval (set to "days" by default).}
}
\value{
Returns a dataframe that has columns for simple moving average and exponential moving average based on specified parameters.
}
\description{
ma creates simple moving averages and exponential moving averages based on parameters specified by the user. It can handle date variables.
}
