% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfit.R
\name{myfit}
\alias{myfit}
\title{myfit}
\usage{
myfit(
  df,
  w,
  y,
  fitobj,
  n,
  mobs = 2,
  interest = c(),
  col = "red",
  linet = "dashed",
  date = FALSE,
  notlm = FALSE
)
}
\arguments{
\item{df}{A dataframe that will be processed.}

\item{w}{Numeric or date vector that will become the x-axis for the final plot.}

\item{y}{Numeric response vector that is the same length as d.}

\item{fitobj}{Fitted object for modeling data.}

\item{n}{Numeric vector containing the desired changepoints to be predicted.}

\item{mobs}{Minimum number of observations between changepoints.}

\item{interest}{Vector containing any points of interest; these points of interest should be on the same scale as d.}

\item{col}{Color of line to indicate points of interest (default set to "red").}

\item{linet}{Type of vertical line to indicate points of interest (default set to "dashed").}

\item{date}{Boolean indicating if the points of interest are dates (default set to FALSE).}

\item{notlm}{Boolean indicating if the model needed is a linear model (default set to FALSE).}
}
\value{
Returns a plotly interactive graph that shows the projected changepoint predictions.
}
\description{
myfit plots user prespecified changepoint predictions over the original data and produces an interactive plotly graph
based on a user-specified fitted object.
}
