% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictsegs.R
\name{predictsegs}
\alias{predictsegs}
\title{predictsegs}
\usage{
predictsegs(
  df,
  w,
  y,
  k,
  z = "bic",
  g = 1,
  interest = c(),
  col = "red",
  linet = "dashed",
  date = FALSE,
  ar = FALSE,
  p = 1,
  d = 0,
  q = 0
)
}
\arguments{
\item{df}{A dataframe that will be processed.}

\item{w}{Numeric or date vector that will become the x-axis for the final plot.}

\item{y}{Numeric response vector that is the same length as w.}

\item{k}{Numeric vector containing the max number of automatically-detected changepoints to be predicted.}

\item{z}{Method used for detecting number of changepoints in the plot (default set to "bic" but "aic" can be used)}

\item{g}{Number of user-specified sub-intervals to examine when searching for changepoints (default set to 1).}

\item{interest}{Vector containing any points of interest; these points of interest should be on the same scale as w.}

\item{col}{Color of line to indicate points of interest (default set to "red").}

\item{linet}{Type of vertical line to indicate points of interest (default set to "dashed").}

\item{date}{Boolean indicating if the points of interest are dates (default set to FALSE).}
}
\value{
Returns a plotly interactive graph that shows the projected changepoint predictions.
}
\description{
predictsegs automatically detects changepoints in the data and returns an interactive plot with the predicted number of changepoints.
Users specify the maximum number of changepoints possible and number of sub-intervals in the data.
}
