% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qclean.R
\name{qclean}
\alias{qclean}
\title{qclean}
\usage{
qclean(df, x, y, a = 0, b = 1, date = TRUE)
}
\arguments{
\item{df}{A dataframe that will be processed.}

\item{x}{The variable to aggregate y by.}

\item{y}{The variable to be aggregated.}

\item{a}{Lower-bound percentile of observations to be removed (set to 0 by default).}

\item{b}{Upper-bound percentile of observations to be removed (set to 1 by default).}

\item{date}{If date is TRUE, converts x variable to POSIXct for numerical analysis (set to TRUE by default).}
}
\value{
Returns a dataframe that is ordered and aggregated by x with percentiles below a and above b removed.
}
\description{
qclean converts date variables to POSIXct, aggregates and orders data by a specified variable, removes specified quantiles,
and creates an index column for later kernel or indexing purposes.
}
