% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qreg.R
\name{qreg}
\alias{qreg}
\title{qreg}
\usage{
qreg(df, x, y, tau = 0.5, intercept = TRUE, m = "fn")
}
\arguments{
\item{df}{A dataframe that will be processed.}

\item{x}{The predictor variable for quantile regression; this can be a vector of strings.}

\item{y}{The response variable as a string for quantile regression.}

\item{tau}{The quantile to be estimate (set to 0.5 by default)}

\item{intercept}{Determines if there is an intercept in the model (set to TRUE by default).}
}
\value{
Returns a fitted model.
}
\description{
qreg creates a quantile regression model using specified predictor and response variables.
}
