% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantnorm.R
\name{quantile_normalize}
\alias{quantile_normalize}
\title{quantnorm}
\usage{
quantile_normalize(df, variable_name)
}
\arguments{
\item{df}{A dataframe that will be processed.}

\item{x}{The variable that will undergo quantile normalization}
}
\value{
Returns a dataframe with an additional column with the normalized values.
}
\description{
quantnorm applies quantile normalization to a variable and returns the dataframe with a column of normalized values.
}
