% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmvpt.R
\name{rmvpt}
\alias{rmvpt}
\title{rmvpt}
\usage{
rmvpt(df, x, y)
}
\arguments{
\item{df}{A dataframe that will be processed.}

\item{x}{Independent variable name as a string for the plot.}

\item{y}{Dependent variable as a string for the plot.}
}
\value{
Returns a dataframe with an additional column with the normalized values.
}
\description{
rmvpt allows users to remove points by clicking on them on an interactive graph.
}
